pyproject.toml
[tool.poetry]
name = "seu-projeto"
version = "0.1.0"
description = "Descrição do seu projeto"
authors = ["Seu Nome <seu.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.13.2"
amqp = "5.2.0"
annotated-types = "0.7.0"
anyio = "4.4.0"
bcrypt = "4.2.0"
billiard = "4.2.0"
cachetools = "5.4.0"
celery = "5.4.0"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
distro = "1.9.0"
dnspython = "2.6.1"
email_validator = "2.2.0"
fastapi = "0.111.1"
fastapi-cli = "0.0.4"
google-ai-generativelanguage = "0.6.6"
google-api-core = "2.19.1"
google-api-python-client = "2.140.0"
google-auth = "2.33.0"
google-auth-httplib2 = "0.2.0"
google-generativeai = "0.7.2"
googleapis-common-protos = "1.63.2"
greenlet = "3.0.3"
grpcio = "1.65.4"
grpcio-status = "1.62.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httplib2 = "0.22.0"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
Jinja2 = "3.1.4"
jiter = "0.5.0"
kombu = "5.3.7"
Mako = "1.3.5"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
mdurl = "0.1.2"
numpy = "2.0.1"
openai = "1.40.3"
pgvector = "0.3.2"
prompt_toolkit = "3.0.47"
proto-plus = "1.24.0"
protobuf = "4.25.4"
psycopg2 = "2.9.9"
pyasn1 = "0.6.0"
pyasn1_modules = "0.4.0"
pydantic = "2.8.2"
pydantic_core = "2.20.1"
Pygments = "2.18.0"
PyJWT = "2.9.0"
pyparsing = "3.1.2"
pypdf = "4.3.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
PyYAML = "6.0.1"
regex = "2024.7.24"
requests = "2.32.3"
rich = "13.7.1"
rsa = "4.9"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
SQLAlchemy = "2.0.31"
SQLAlchemy-Utils = "0.41.2"
starlette = "0.37.2"
tiktoken = "0.7.0"
tqdm = "4.66.5"
typer = "0.12.3"
typing_extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.2"
uvicorn = "0.30.3"
vine = "5.1.0"
watchfiles = "0.22.0"
wcwidth = "0.2.13"
websockets = "12.0"

[tool.poetry.dev-dependencies]
# Adicione suas dependências de desenvolvimento aqui

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
